refactor: create nonop & op command file
