+++ src/generate/utils/nonopininated-cmd.ts
@@ -1,0 +1,4 @@
+export async function nonOpinionatedProcess(): Promise<string> {
+	// Non opinionated
+	return "";
+}
